Refactoring. Changed names, descriptions and float fields to value objects. Added FluentValidation for CreateVolunteerCommand validation.

+++ PetFamily.Backend/src/PetFamily.Application/Volunteers/DTOs/VolunteerDTO.cs
@@ -5,7 +5,7 @@ public record VolunteerDTO(string FirstName,
     string FatherName,
     string Email,
     string Description,
-    double ExperienceYears,
+    float ExperienceYears,
     string Phone)
 {
